*RankingGenre/RankingMain cleanup logic *RankingGenre derived BigButton to GenreBigButton *general cleanup
