+++ src/scenes/session/Login/LoginForm.tsx
@@ -88,15 +88,6 @@ export default class LoginForm extends Component<LoginFormProps, LoginFormState>
                         variant='contained'
                         color='primary'
                     />
-                    {/*TODO: Uncomment in password recovery*/}
-                    {/*<Button*/}
-                    {/*    title='No sé mi contraseña'*/}
-                    {/*    disabled={false}*/}
-                    {/*    loading={loading}*/}
-                    {/*    onClick={this.props.onPasswordRecover}*/}
-                    {/*    variant='text'*/}
-                    {/*    color='secondary'*/}
-                    {/*/>*/}
                 </div>
             </div>
         )
